requirements.txt
uvicorn[standard]
pydantic~=2.11.2
fastapi~=0.115.12
sqlalchemy~=2.0.12
databases[asyncpg]~=0.9.0
python-dotenv~=1.0.0
pydantic-settings~=2.8.1
python-dotenv
rich
asgi-correlation-id~=4.3.4
python-json-logger
python-multipart
python-jose~=3.3.0
passlib[bcrypt]~=1.7.4



requests~=2.32.3
pytest~=8.3.5
httpx~=0.28.1
alembic~=1.10.4